review: blocking fallback when no thread can start, guard rewatch, unref with the process's own loop, move the test to terminal.test.ts
